UI：CustomButton を親クラスに設定・CommonDesignButton を作成
