+++ Design/source/CustomButton.swift
@@ -9,7 +9,7 @@
 import UIKit
 import Model
 
-@IBDesignable public final class CustomButton: UIButton {
+@IBDesignable public class CustomButton: UIButton {
     
     public var colorSet: ColorSet = .default {
         didSet {
@@ -27,35 +27,9 @@ import Model
         configure()
     }
     
-    private func configure() {
+    func configure() {
         setColorSet()
-        setAlpha()
     }
     
-    private func setColorSet() {
-        backgroundColor = colorSet.navigationBarColor()
-        layer.cornerRadius = 5
-        
-        guard let text = titleLabel?.text else { return }
-        let attributedText = NSAttributedString(string: text,
-                                                        attributes: [.font : UIFont.boldSystemFont(ofSize: 15),
-                                                                     .foregroundColor : colorSet.backgroundColor()])
-        setAttributedTitle(attributedText, for: .normal)
-    }
-    
-    public override var intrinsicContentSize: CGSize {
-        let margin: CGFloat = 40
-        let titleWidth = titleLabel?.intrinsicContentSize.width ?? 0
-        return CGSize(width: titleWidth + margin, height: 30)
-    }
-    
-    override public var isEnabled: Bool {
-        didSet {
-            setAlpha()
-        }
-    }
-    
-    private func setAlpha() {
-        alpha = isEnabled ? 1 : 0.3
-    }
+    func setColorSet() {}
 }
